Export TDigest as a struct, not an interface
Good Go APIs use concrete types, not interfaces, as their exported
values. This makes it easier to add methods or fields in the future.
There was no benefit to this package's earlier use of an interface
type; the only code that accepted a TDigest as input had a hidden
requirement that the TDigest be a *centroidSet anyway.