Re-type TDigest as a struct, not an interface
Exporting an interface is awkward. It makes it hard to safely add new
methods to the exported type. There was no gain in usability from
using an interface, in this case, either - only one method, MergeInto,
actually accepted a TDigest interface as input - and even then it
secretly required that the input be a *centroidSet!

Instead, we should export a struct. This is an API-breaking change,
but it's worthwhile.