Drop 'compatibility' with reference implementation's serialization
The reference implementation stores values we don't use and doesn't
have a documented spec. It uses two different encodings, but we only
need one. It also uses float weights, while we use integers.

Compatibility was a bit of a sham before anyway, so it's best to drop
the charade and use our own scheme.